move upsert and a handful of helpers to InMemAccountsIndex
